# Conflicts:
#	th2_data_services/data.py

+++ th2_data_services/data.py
@@ -1,3 +1,17 @@
+#  Copyright 2022 Exactpro (Exactpro Systems Limited)
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
 import copy
 import pickle
 import pprint
@@ -26,7 +40,7 @@ class Data:
             data: Data source.
             workflow: Workflow.
             parents_cache: Parents chain. Works as a stack.
-            cache: Flag if you want write and read from cache.
+            cache: Flag if you want to write and read from cache.
         """
         self._cache_filename = f"{str(id(self))}:{time()}.pickle"
         self._len = None
@@ -49,7 +63,8 @@ class Data:
     def __delete_cache(self) -> None:
         """Removes cache file."""
         path = Path(self.__get_cache_filepath())
-        path.unlink()
+        if path.exists():
+            path.unlink()
 
     @property
     def len(self) -> int:
@@ -113,7 +128,7 @@ class Data:
         """Loads data from cache or data.
 
         Args:
-            cache: Flag if you what write and read from cache.
+            cache: Flag if you what to write and read from cache.
 
         Returns:
             obj: Generator
@@ -133,8 +148,6 @@ class Data:
             if self.__check_file_recording():
                 # Do not read from the cache file if it has PENDING status (if the file is not filled yet).
                 cache = False
-
-            workflow = self._build_workflow(workflow)
 
             yield from self.__change_data(working_data=working_data, workflow=workflow, cache=cache)
 
@@ -364,7 +377,6 @@ class Data:
         new_parents_cache = [*self._parents_cache, self._cache_filename]
         data_obj = Data(data=self._data, workflow=new_workflow, parents_cache=new_parents_cache)
         data_obj._length_hint = num
-        data_obj._limit_num = num
         return data_obj
 
     def sift(self, limit: int = None, skip: int = None) -> Generator[dict, None, None]:
@@ -394,7 +406,7 @@ class Data:
         """Change status cache.
 
         If True all requested data from rpt-data-provider will be saved to cache file.
-        Further actions with Data object will be consume data from the cache file.
+        Further actions with the Data object will consume data from the cache file.
 
         Args:
             status(bool): Status.
